Add tests for all engines
