+++ ktor-server/ktor-server-netty/build.gradle.kts
@@ -40,7 +40,6 @@ kotlin.sourceSets {
             api(libs.logback.classic)
 
             api(project(":ktor-server:ktor-server-core", configuration = "testOutput"))
-            implementation(project(":ktor-server:ktor-server-plugins:ktor-server-hsts"))
         }
     }
 }
